+++ components/wishes/[id].tsx
